refactor: PostViewController의 Extension을 개별 파일로 분리
