+++ OpenMarket/OpenMarket/View Controller/Post/PostViewController.swift
@@ -62,126 +62,6 @@ extension PostViewController: UICollectionViewDelegateFlowLayout {
     }
 }
 
-// MARK: Bar Button Action(Cancel, Post)
-extension PostViewController {
-    @IBAction func backToListView(_ sender: Any) {
-        self.dismiss(animated: true, completion: nil)
-    }
-    
-    @IBAction func postProduct(_ sender: Any) {
-        let checkValidate: Result<ProductParams, ViewControllerError> = checkValidData()
-        let urlSessionProvider = URLSessionProvider()
-        
-        // remove PlusButtonImage
-        if tryAddImageCount < 5 {
-            images.removeLast()
-        }
-        
-        switch checkValidate {
-        case .success(let data):
-            urlSessionProvider.postData(requestType: .productRegistration, params: data, images: self.images) { (result: Result<Data, NetworkError>) in
-                switch result {
-                case .success(_):
-                    self.goToPreviousView(title: "업로드 성공", "제품 리스트 화면으로 이동합니다")
-                case .failure(let error):
-                    self.alertError(error)
-                }
-            }
-        case .failure(let error):
-            self.alertError(error)
-        }
-    }
-    
-    func checkValidData() -> Result<ProductParams, ViewControllerError> {
-        guard let name = productNameTextField.text else {
-            return .failure(.invalidInput)
-        }
-        
-        guard name.count >= 3 else {
-            return .failure(.invalidInput)
-        }
-        
-        guard let priceTextfieldText = priceTextField.text else {
-            return .failure(.invalidInput)
-        }
-        
-        guard let price = Double(priceTextfieldText) else {
-            return .failure(.invalidInput)
-        }
-        
-        var currency = Currency.KRW
-        let currencyIndex = currencySelectButton.selectedSegmentIndex
-        switch currencyIndex {
-        case 0:
-            currency = Currency.KRW
-        case 1:
-            currency = Currency.USD
-        default:
-            return .failure(.invalidInput)
-        }
-        
-        guard let discount = Double(bargainPriceTextField.text ?? "0") else {
-            return .failure(.invalidInput)
-        }
-        
-        let bargainPrice = price - discount
-        
-        guard let stock = Int(stockTextField.text ?? "0") else {
-            return .failure(.invalidInput)
-        }
-        
-        guard let description = descriptionTextView.text else {
-            return .failure(.invalidInput)
-        }
-        
-        return .success(ProductParams(name: name,
-                                      descriptions: description,
-                                      price: price,
-                                      currency: currency,
-                                      discountedPrice: bargainPrice,
-                                      stock: stock,
-                                      secret: "c8%MC*3wjXJ?Wf+g"))
-    }
-}
-
-// MARK: Text View Delegate
-extension PostViewController: UITextViewDelegate {
-    @objc
-    func keyboardWillShow(_ sender: Notification) {
-        self.view.frame.origin.y = -150
-    }
-    
-    @objc
-    func keyboardWillHide(_ sender: Notification) {
-        self.view.frame.origin.y = 0
-    }
-    
-    func textViewDidEndEditing(_ textView: UITextView) {
-        if self.descriptionTextView.text.isEmpty {
-            descriptionTextView.text = textViewPlaceHolder
-            descriptionTextView.textColor = .lightGray
-        }
-    }
-    
-    func textViewDidBeginEditing(_ textView: UITextView) {
-        if self.descriptionTextView.text == textViewPlaceHolder {
-            descriptionTextView.text = nil
-            descriptionTextView.textColor = .black
-        }
-    }
-    
-    func textView(_ textView: UITextView, shouldChangeTextIn range: NSRange, replacementText text: String) -> Bool {
-        let currentText = textView.text ?? ""
-        guard let stringRange = Range(range, in: currentText) else {
-            return false
-        }
-        
-        let changedText = currentText.replacingCharacters(in: stringRange, with: text)
-        
-        return changedText.count <= 1000
-    }
-}
-
 // MARK: Collection View Data Source, Delegation
 extension PostViewController: UICollectionViewDataSource, UICollectionViewDelegate {
     func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
